BIS/requirements.txt
django_filter==2.3.0
plotly==4.9.0
pivottablejs==0.9.0
django-phonenumber-field
XlsxWriter==1.3.7
django_ckeditor==6.0.0
Django==3.1
seaborn==0.10.1
matplotlib==3.3.4
xhtml2pdf==0.2.4
django_debug_toolbar==3.2
six==1.15.0
reportlab==3.6.6
whitenoise
psycopg2-binary
ipython
django-crispy-forms
openpyxl
docutils
xlsx-streaming
